backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
psycopg2-binary
pydantic
python-dotenv
openai
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
